Khang

+++ src/main/java/com/springboot/wmproject/DTO/FoodDetailDTO.java
@@ -4,8 +4,7 @@ import com.springboot.wmproject.entities.Food;
 import com.springboot.wmproject.entities.Orders;
 import lombok.Data;
 
-import javax.persistence.*;
-import java.math.BigDecimal;
+
 @Data
 public class FoodDetailDTO {
     private int id;
@@ -16,7 +15,7 @@ public class FoodDetailDTO {
 
     private Integer count;
 
-    private BigDecimal price;
+    private Double price;
 
     private Orders ordersByOrderId;
     private Food foodByFoodId;
